GH-3698: Allow JSON-LD conversion outside of read/write operations
